Backend/requirements.txt

flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
PyJWT==2.8.0
Werkzeug==2.3.6
mysql-connector-python==8.1.0
pymysql==1.1.0